+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Acquisition - New acquisition protocol for thresholded acquisitions: `ThresholdedAcquisition` (!617)
 - Pin version of `dataclasses-json` due to breaking pipelines. (!727)
 - Schedulables - Store references to `Schedulables` in timing contraints as `string` rather than `Schedulable`. (!717)
+- Qblox backend - Fix weighted acquisition in append mode. (!725)
 
 ## 0.15.0 (2023-07-13)